separate outfit generation and graph operations and add requirement for s3
