+++ api/closet.py
@@ -2,6 +2,7 @@ from flask import jsonify
 from data_access.closet_dao import closet_dao
 from model.closet_model import Closet
 from ml.outfit_generator import get_top_outfits
+from ml.graph_manager import generate_graph
 
 def create_closet(username, closet):
     """
@@ -63,12 +64,37 @@ def get_closet(username, closet_name):
 # TODO: make outfit_images an input instead of hardcode
 # TODO: move get_best_outfit to closet_dao
 def get_best_outfit():
-    # run get_outfit_score along with the algorithm
     try:
-        # TODO: add algorithm, just returning one score for testing purposes
-        # complete path to the images
-        # outfit_images = ['test-outfit/top.jpg','test-outfit/bottom.jpg','test-outfit/shoes.jpg','test-outfit/bag.jpg','test-outfit/accessory.jpg']
-        score_returned = get_top_outfits(['test-outfit/top.jpg'],['test-outfit/bottom.jpg'],['test-outfit/shoes.jpg'],['test-outfit/bag.jpg'],['test-outfit/accessory.jpg'])
+        # hardcode the clothes dictionary for now, will retrieve from s3 and process in the future
+        # this is a dictionary with the list of item names used to create nodes in the graph
+        clothes = {'tops':['test-outfit/top.jpg'],
+        'bottoms':['test-outfit/bottom.jpg'],
+        'shoes': ['test-outfit/shoes.jpg'],
+        'bags': ['test-outfit/bag.jpg'],
+        'accessories':['test-outfit/accessory.jpg']}
+
+        # TODO: change this to fetch closet graph instead of creating
+        graph = create_closet_graph()
+        
+        score_returned = get_top_outfits(graph, clothes)
         return jsonify(score = score_returned)
     except Exception as error:
         return jsonify(error = str(error)), 500
+
+def create_closet_graph():
+    # TODO: get clothings from db or pass it in instead of hardcode
+    graph = generate_graph(['test-outfit/top.jpg'],['test-outfit/bottom.jpg'],['test-outfit/shoes.jpg'],['test-outfit/bag.jpg'],['test-outfit/accessory.jpg'])
+    # TODO: save this graph to s3
+    return graph
+
+def get_closet_graph():
+    # TODO: some s3 connection that fetches graph from db
+    pass
+
+def add_clothes_to_closet():
+    # TODO: get the graph and perform graph operation using graph manager
+    pass
+
+def remove_clothes_from_closet():
+    # TODO: get the graph and perform graph operation using graph manager
+    pass
