+++ app.py
@@ -34,6 +34,6 @@ if __name__ == '__main__':
         sqla.create_all()
 
     try:
-        app.run(host='0.0.0.0', port=int(os.getenv('PORT', 5000)), debug=True)
+        app.run(host='0.0.0.0', port=int(os.getenv('PORT', 8080)), debug=True)
     finally:
         clean_up()
