+++ requirements.txt
@@ -36,3 +36,4 @@ torch==1.4.0
 torchvision==0.2.1
 zipp==3.4.0
 importlib-metadata==3.4.0
+Flask-SQLAlchemy==2.4.4
